Moved two classes in another package
